requirements.txt
flask
flask-cors
flask-login
python-dotenv
pymongo